Fix lib_v_fillarea() stack smash
When calling C functions, the parameters pushed on the stack are the
property of the callee. It can't be assumed that they will not be changed
when the function returns. And this is precisely what happens with GCC
4.6.4 when lib_v_fillarea() calls filled_poly(): the latter calls
fill_spans() with a simple jmp, after having reorganized the stack
contents. So the argument passed to free_block() was wrong.

This commit fixes Kronos VDI benchmark when v_fillarea() is not
accelerated.